drivers: host: scsi: Avoid integer overflow
Expression evaluated using 32-bit arithmetic can overflow the result
which is subsequently used for comparison with variable of type u64.
Cast the expression to u64 to avoid unexpected behaviour.

Coverity ID 29614

Bug 200192581

Reviewed-on: http://git-master/r/1324206
(cherry picked from commit 12452234a29def3a4f170d5476a024d50e6325c6)
Reviewed-on: https://git-master.nvidia.com/r/1528635
GVS: Gerrit_Virtual_Submit
Tested-by: Rakesh Krishna Chekuri <rchekuri@nvidia.com>

+++ drivers/nvme/host/scsi.c
@@ -1761,7 +1761,7 @@ static int nvme_trans_io(struct nvme_ns *ns, struct sg_io_hdr *hdr, u8 is_write,
 	xfer_bytes = min(((u64)hdr->dxfer_len), sum_iov_len);
 
 	/* If block count and actual data buffer size dont match, error out */
-	if (xfer_bytes != (cdb_info.xfer_len << ns->lba_shift)) {
+	if (xfer_bytes != ((u64)cdb_info.xfer_len << ns->lba_shift)) {
 		res = -EINVAL;
 		goto out;
 	}
